refactor: Create some default implementations for Menu methods.
