+++ Assets/UnZENity-Core/Scripts/UI/Menus/MainMenu.cs
@@ -23,11 +23,6 @@ namespace GUZ.Core.UI.Menus
             return;
         }
 
-        protected override void Back(string itemName, string commandName)
-        {
-            MenuHandler.BackMenu();
-        }
-
         protected override void StartMenu(string itemName, string commandName)
         {
             MenuHandler.OpenMenu(commandName);
