+++ Assets/UnZENity-Core/Scripts/UI/Menus/QuestLogMenu.cs
@@ -407,11 +407,6 @@ namespace GUZ.Core.UI.Menus
             throw new NotImplementedException();
         }
 
-        protected override void Back(string itemName, string commandName)
-        {
-            throw new NotImplementedException();
-        }
-
         protected override void StartMenu(string itemName, string commandName)
         {
             throw new NotImplementedException();
