+++ Assets/UnZENity-Core/Scripts/UI/Menus/SettingMenu.cs
@@ -65,11 +65,6 @@ namespace GUZ.Core.UI.Menus
             throw new NotImplementedException();
         }
 
-        protected override void Back(string itemName, string commandName)
-        {
-            MenuHandler.BackMenu();
-        }
-
         protected override void StartMenu(string itemName, string commandName)
         {
             throw new NotImplementedException();
@@ -96,11 +91,6 @@ namespace GUZ.Core.UI.Menus
         }
 
         protected override void ExecuteCommand(string itemName, string commandName)
-        {
-            throw new NotImplementedException();
-        }
-
-        protected override bool IsMenuItemInitiallyActive(string menuItemName)
         {
             throw new NotImplementedException();
         }
